canjes devoluciones 85%

+++ app/src/main/java/union/union_vr1/Sqlite/DBAdapter_Temp_Canjes_Devoluciones.java
@@ -7,6 +7,12 @@ import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
 
+import java.text.DecimalFormat;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
+import java.util.GregorianCalendar;
+
 import union.union_vr1.Conexion.DbHelper;
 
 /**
@@ -30,10 +36,12 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     public static final String temp_id_documento = "temp_id_documento";
     public static final String temp_id_motivo = "temp_id_motivo";
     public static final String temp_estado_producto = "temp_id_estado_producto";
+    public static final String temp_liquidacion="temp_liquidacion";
+    public static final String temp_id_detalle = "temp_id_detalle";
 
     public static final String temp_estado = "temp_estado";
 
-
+    public DecimalFormat f = new DecimalFormat("##.00");
     public static final String TAG = "Temp_Venta_Detalle";
 
     private DbHelper mDbHelper;
@@ -46,20 +54,22 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     public static final String CREATE_TABLE_TEMP_CANJES_DEVOLUCIONES =
             "create table " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " ("
                     + temp_id_canjes_devoluciones + " integer primary key autoincrement,"
-                    + temp_id_comprob+" text,"
-                    + temp_id_producto+" text,"
-                    + temp_nom_producto+" text,"
-                    + temp_codigo_producto+" text,"
-                    + temp_cantidad+" text,"
-                    + temp_precio_unit+ " text,"
-                    + temp_importe+" text,"
-                    + temp_lote+" text,"
-                    + temp_fecha_emision+" text,"
-                    + temp_cliente+" text,"
-                    + temp_id_establecimiento+" text,"
-                    + temp_id_documento+" text,"
-                    + temp_id_motivo+" text,"
-                    + temp_estado_producto+" text,"
+                    + temp_id_comprob + " text,"
+                    + temp_id_producto + " text,"
+                    + temp_nom_producto + " text,"
+                    + temp_codigo_producto + " text,"
+                    + temp_cantidad + " text,"
+                    + temp_precio_unit + " text,"
+                    + temp_importe + " text,"
+                    + temp_lote + " text,"
+                    + temp_fecha_emision + " text,"
+                    + temp_cliente + " text,"
+                    + temp_id_establecimiento + " text,"
+                    + temp_id_documento + " text,"
+                    + temp_id_motivo + " text,"
+                    + temp_estado_producto + " text,"
+                    +temp_id_detalle +" text,"
+                    +temp_liquidacion+ " text,"
                     + temp_estado + " integer);";
 
 
@@ -82,7 +92,7 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
     }
 
     public long createTempCanjesDevoluciones(String id_comprob, String id_producto, String nom_producto, String codigo_producto, String cantidad, String precio_unit, String importe, String lote, String fecha_emision, String cliente,
-                                       String id_establecimiento, String id_documento, String id_motivo, String estado_producto, String estado) {
+                                             String id_establecimiento, String id_documento, String id_motivo, String estado_producto, String estado,String liquidacion,String idDetalle) {
 
         ContentValues initialValues = new ContentValues();
         initialValues.put(temp_id_comprob, id_comprob);
@@ -100,16 +110,78 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
         initialValues.put(temp_id_motivo, id_motivo);
         initialValues.put(temp_estado_producto, estado_producto);
         initialValues.put(temp_estado, estado);
+        initialValues.put(temp_liquidacion, liquidacion);
+        initialValues.put(temp_id_detalle,idDetalle);
+        initialValues.put(temp_fecha_emision, getDatePhone());
 
         return mDb.insert(SQLITE_TABLE_Temp_Canjes_Devoluciones, null, initialValues);
     }
-    public Cursor listarCanjes(){
-        Cursor cursor = mDb.rawQuery("select * from "+SQLITE_TABLE_Temp_Canjes_Devoluciones+" where "+temp_id_motivo+"='2';",null);
-        return  cursor;
+
+
+    public Cursor getAllOperacion (String fecha,String establec){
+        Cursor cursor = mDb.rawQuery("select * from "+SQLITE_TABLE_Temp_Canjes_Devoluciones+" where "+temp_fecha_emision+"='"+fecha+"' and "+temp_id_establecimiento+"='"+establec+"'",null);
+        return cursor;
     }
-    public Cursor listarDevoluciones(){
-        Cursor cursor = mDb.rawQuery("select * from "+SQLITE_TABLE_Temp_Canjes_Devoluciones+" where "+temp_id_motivo+"='1';",null);
-        return  cursor;
+
+    public Cursor listarCanjes(String establec) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='2' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "';", null);
+        return cursor;
+    }
+
+    public String[] getInforHeader(String idEstablecimiento) {
+        Cursor cabecera;
+        cabecera = mDb.rawQuery("select * from m_evento_establec where ee_in_id_establec ='" + idEstablecimiento + "'", null);
+        String cliente = "", documento = "";
+        if (cabecera.moveToFirst()) {
+            cliente = cabecera.getString(cabecera.getColumnIndexOrThrow(DbAdaptert_Evento_Establec.EE_nom_cliente));
+            documento = cabecera.getString(cabecera.getColumnIndexOrThrow(DbAdaptert_Evento_Establec.EE_doc_cliente));
+        }
+        return new String[]{cliente, documento};
+    }
+
+    public Double[] getInfoCanje(String establec) {
+        Cursor cursor = mDb.rawQuery("select sum(" + temp_importe + ") as " + temp_importe + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='2' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "';", null);
+        double total = 0.0000;
+        double igv = 0.0000;
+        double base = 0.0000;
+        cursor.moveToFirst();
+        if (cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe)) > 0) {
+            total = cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe));
+            base = total / 1.18;
+            igv = total - base;
+            Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base));
+        }
+        return new Double[]{Double.parseDouble(f.format(total)), Double.parseDouble(f.format(base)), Double.parseDouble(f.format(igv))};
+    }
+
+    private String getDatePhone() {
+        Calendar cal = new GregorianCalendar();
+        Date date = cal.getTime();
+        SimpleDateFormat df = new SimpleDateFormat("dd/MM/yyyy");
+        String formatteDate = df.format(date);
+        return formatteDate;
+    }
+
+    public Cursor listarDevoluciones(String establec) {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='1' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "';", null);
+        return cursor;
+    }
+
+    public Double[] getInfoDevoluciones(String establec) {
+        double total = 0.0000;
+        double igv = 0.00000;
+        double base = 0.00000;
+        Cursor cursor = mDb.rawQuery("select sum(" + temp_importe + ") as " + temp_importe + " from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + " where " + temp_id_motivo + "='1' and " + temp_id_establecimiento + "='" + establec + "' and " + temp_fecha_emision + "='" + getDatePhone() + "';", null);
+        cursor.moveToFirst();
+
+        if (cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe)) > 0) {
+            total = cursor.getDouble(cursor.getColumnIndexOrThrow(temp_importe));
+            base = total / 1.18;
+            igv = total - base;
+            Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base));
+        }
+        Log.d("INFO", ":Total: " + f.format(total) + " Igv: " + f.format(igv) + " BASE: " + f.format(base));
+        return new Double[]{Double.parseDouble(f.format(total)), Double.parseDouble(f.format(base)), Double.parseDouble(f.format(igv))};
     }
 
     public void updateTempCanjesDevoluciones2(String idorig, String iddest, String vals) {
@@ -130,13 +202,20 @@ public class DBAdapter_Temp_Canjes_Devoluciones {
 
 
     }
-    public int truncateCanjesDevoluciones(){
-        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones,null,null);
-    }
-    public int deleteCanjesDevoluciones(int id){
-        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, temp_id_canjes_devoluciones + " = ?", new String[] { id+"" });
+
+    public int truncateCanjesDevoluciones() {
+        Cursor cursor = mDb.rawQuery("select * from " + SQLITE_TABLE_Temp_Canjes_Devoluciones + "", null);
+        if (cursor.moveToFirst()) {
+            return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, null, null);
+        } else {
+            return 1;
+        }
+
     }
 
+    public int deleteCanjesDevoluciones(int id) {
+        return mDb.delete(SQLITE_TABLE_Temp_Canjes_Devoluciones, temp_id_canjes_devoluciones + " = ?", new String[]{id + ""});
+    }
 
 
 }
